Consolidate tests into fewer binaries
Running our test suite is generally a fast thing, but velocity is a
relative thing. The slowest part when running all tests comes from
the overhead of running the various test binaries. In large part that
seems to be due to the default test runner, with cargo-nextest
supposedly showing how things can be sped up -- except that for us it
causes an ~80x slow down.
In an attempt to cut down the number of test binaries we have, integrate
the permissionless and permission tests into the "integration" test
suite. In order to go about that, use the newly introduced test-fork
crate, which can be used to run individual tests as separate processes.
