+++ Cargo.toml
@@ -156,6 +156,7 @@ rand = {version = "0.9", default-features = false, features = ["std", "thread_rn
 scopeguard = "1.2"
 stats_alloc = {version = "0.1.1", features = ["nightly"]}
 tempfile = "3.4"
+test-fork = "0.1"
 test-log = {version = "0.2.14", default-features = false, features = ["trace"]}
 test-tag = "0.1.3"
