+++ tests/suite/mod.rs
@@ -1,3 +1,4 @@
+mod common;
 mod inspect;
 mod normalize;
 mod symbolize;
